Restructure authentication code
The original structure used an `Account` class that contained
interactions for both the yggdrasil system and the new xbox live
(microsoft) system.

This change isolates these into their own files, with each system having
their own Account class, being a subclass of a new base `Account` class.
